refactor: renderer native utils
