+++ src/renderer/utils/native.js
@@ -150,3 +150,17 @@ export const openExternal = (url, options) => {
 
   remote.shell.openExternal(url, options)
 }
+
+export const delayDeleteTaskFiles = (task, delay) => {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      try {
+        const result = moveTaskFilesToTrash(task)
+        resolve(result)
+      } catch (err) {
+        console.log('[Motrix] batch delay delete task files fail', err)
+        resolve(false)
+      }
+    }, delay)
+  })
+}
